o use SCM commands rather than manually building them o don't assertDirectory for empty - checkout does a prune in local provider
git-svn-id: https://svn.apache.org/repos/asf/maven/scm/trunk@162244 13f79535-47bb-0310-9956-ffa450edef68

+++ maven-scm-providers/maven-scm-provider-local/src/main/java/org/apache/maven/scm/provider/local/command/checkout/LocalCheckOutCommand.java
@@ -61,11 +61,6 @@ public class LocalCheckOutCommand
 
         File baseDestination = fileSet.getBasedir();
 
-        if ( !baseDestination.exists() )
-        {
-            throw new ScmException( "The working directory doesn't exist (" + baseDestination.getAbsolutePath() + ")." );
-        }
-
         if ( !root.exists() )
         {
             throw new ScmException( "The base directory doesn't exist (" + root.getAbsolutePath() + ")." );
@@ -80,7 +75,10 @@ public class LocalCheckOutCommand
 
         try
         {
-        	FileUtils.deleteDirectory( baseDestination );
+            if ( baseDestination.exists() )
+            {
+        	    FileUtils.deleteDirectory( baseDestination );
+            }
 
         	if ( !baseDestination.mkdirs() )
         	{
